Added snapshot testing
Using snapshot testing because the underlying HTML generated by the
components should never change. If we change our data, the underlying
structure should not change! When our data changes, the only thing that
should change are the HTML attributes which correspond to the data
passed in.

As we refactor and change our logic throughout our componets, this
should not affect our tests! If our snapshot tests fail, that means we
did something wrong!